api_rest/requirements.txt
django == 5.2
djangorestframework == 3.16.0
django-cors-headers == 4.7.0